+++ src/QuickPID.h
@@ -19,15 +19,15 @@ class QuickPID {
     QuickPID();
 
     // Constructor. Links the PID to Input, Output, Setpoint, initial tuning parameters and control modes.
-    QuickPID(double *Input, double *Output, double *Setpoint, double Kp, double Ki, double Kd,
+    QuickPID(float *Input, float *Output, float *Setpoint, float Kp, float Ki, float Kd,
              pMode pMode, dMode dMode, iAwMode iAwMode, Action Action);
 
     // Overload constructor links the PID to Input, Output, Setpoint, tuning parameters and control Action.
     // Uses defaults for remaining parameters.
-    QuickPID(double *Input, double *Output, double *Setpoint, double Kp, double Ki, double Kd, Action Action);
+    QuickPID(float *Input, float *Output, float *Setpoint, float Kp, float Ki, float Kd, Action Action);
 
     // Simplified constructor which uses defaults for remaining parameters.
-    QuickPID(double *Input, double *Output, double *Setpoint);
+    QuickPID(float *Input, float *Output, float *Setpoint);
 
     // Sets PID mode to manual (0), automatic (1), timer (2) or toggle manual/automatic (3).
     void SetMode(Control Mode);
@@ -38,16 +38,16 @@ class QuickPID {
     bool Compute();
 
     // Sets and clamps the output to a specific range (0-255 by default).
-    void SetOutputLimits(double Min, double Max);
+    void SetOutputLimits(float Min, float Max);
 
     // available but not commonly used functions ******************************************************************
 
     // While most users will set the tunings once in the constructor, this function gives the user the option of
     // changing tunings during runtime for Adaptive control.
-    void SetTunings(double Kp, double Ki, double Kd);
+    void SetTunings(float Kp, float Ki, float Kd);
 
     // Overload for specifying proportional ratio.
-    void SetTunings(double Kp, double Ki, double Kd, pMode pMode, dMode dMode, iAwMode iAwMode);
+    void SetTunings(float Kp, float Ki, float Kd, pMode pMode, dMode dMode, iAwMode iAwMode);
 
     // Sets the controller direction or action. Direct means the output will increase when the error is positive.
     // Reverse means the output will decrease when the error is positive.
@@ -74,43 +74,43 @@ class QuickPID {
     void SetAntiWindupMode(uint8_t IawMode);
 
     // sets the output summation value
-    void SetOutputSum(double sum);
+    void SetOutputSum(float sum);
 
     void Initialize();        // Ensure a bumpless transfer from manual to automatic mode
     void Reset();             // Clears pTerm, iTerm, dTerm and outputSum values
 
     // PID Query functions ****************************************************************************************
-    double GetKp();            // proportional gain
-    double GetKi();            // integral gain
-    double GetKd();            // derivative gain
-    double GetPterm();         // proportional component of output
-    double GetIterm();         // integral component of output
-    double GetDterm();         // derivative component of output
-    double GetOutputSum();     // summation of all pid term components
+    float GetKp();            // proportional gain
+    float GetKi();            // integral gain
+    float GetKd();            // derivative gain
+    float GetPterm();         // proportional component of output
+    float GetIterm();         // integral component of output
+    float GetDterm();         // derivative component of output
+    float GetOutputSum();     // summation of all pid term components
     uint8_t GetMode();        // manual (0), automatic (1), timer (2) or toggle manual/automatic (3)
     uint8_t GetDirection();   // direct (0), reverse (1)
     uint8_t GetPmode();       // pOnError (0), pOnMeas (1), pOnErrorMeas (2)
     uint8_t GetDmode();       // dOnError (0), dOnMeas (1)
     uint8_t GetAwMode();      // iAwCondition (0, iAwClamp (1), iAwOff (2)
 
-    double outputSum;          // Internal integral sum
+    float outputSum;          // Internal integral sum
 
   private:
 
-    double dispKp = 0;   // for defaults and display
-    double dispKi = 0;
-    double dispKd = 0;
-    double pTerm;
-    double iTerm;
-    double dTerm;
+    float dispKp = 0;   // for defaults and display
+    float dispKi = 0;
+    float dispKd = 0;
+    float pTerm;
+    float iTerm;
+    float dTerm;
 
-    double kp;           // (P)roportional Tuning Parameter
-    double ki;           // (I)ntegral Tuning Parameter
-    double kd;           // (D)erivative Tuning Parameter
+    float kp;           // (P)roportional Tuning Parameter
+    float ki;           // (I)ntegral Tuning Parameter
+    float kd;           // (D)erivative Tuning Parameter
 
-    double *myInput;     // Pointers to the Input, Output, and Setpoint variables. This creates a
-    double *myOutput;    // hard link between the variables and the PID, freeing the user from having
-    double *mySetpoint;  // to constantly tell us what these values are. With pointers we'll just know.
+    float *myInput;     // Pointers to the Input, Output, and Setpoint variables. This creates a
+    float *myOutput;    // hard link between the variables and the PID, freeing the user from having
+    float *mySetpoint;  // to constantly tell us what these values are. With pointers we'll just know.
 
     Control mode = Control::manual;
     Action action = Action::direct;
@@ -119,7 +119,7 @@ class QuickPID {
     iAwMode iawmode = iAwMode::iAwCondition;
 
     uint32_t sampleTimeUs, lastTime;
-    double outMin, outMax, error, lastError, lastInput;
+    float outMin, outMax, error, lastError, lastInput;
 
 }; // class QuickPID
 #endif // QuickPID.h
